Replace bootstrap with react bootstrap
