+++ frontend/src/app/landing/styles/landingPage.scss
@@ -1,5 +1,5 @@
 .team img {
-    width: 100px;
-    height: 100px;
+    width: 7rem;
+    height: 7rem;
     border-radius: 50%;
 }
