+++ frontend/src/app/pages/desktop.homePage.jsx
@@ -1,7 +1,7 @@
 // @flow
 import * as React from 'react';
 
-import LandingPage from '../landing/LandingPage';
+import LandingPage from '../landing/landingPage';
 
 const HomePage = (): React.Node => {
     return (
